# Added Signed Cookie and optionally used it in Session - Other Speedups