--- a/spec/amber/router/session/redis_store_spec.cr
+++ b/spec/amber/router/session/redis_store_spec.cr
@@ -2,8 +2,8 @@ require "../../../../spec_helper"
 require "redis"
 
 module Amber::Router::Session
-  REDIS_STORE = Redis.new(url: ENV["REDIS_URL"] || "redis://127.0.0.1:6379")
-  
+  REDIS_STORE = Redis.new(url: ENV["REDIS_URL"]? || "redis://127.0.0.1:6379")
+
   describe RedisStore do
     describe "#id" do
       it "returns a UUID" do