--- a/src/amber/router/pipe/session.cr
+++ b/src/amber/router/pipe/session.cr
@@ -6,9 +6,11 @@ module Amber
     class Session < Base
       def call(context : HTTP::Server::Context)
         call_next(context)
-        context.session.set_session
-        context.cookies.write(context.response.headers)
-        context
+
+        if context.session.changed?
+          context.session.set_session
+          context.cookies.write(context.response.headers)
+        end
       end
     end
   end