--- a/src/amber/router/session/cookie_store.cr
+++ b/src/amber/router/session/cookie_store.cr
@@ -22,7 +22,7 @@ module Amber::Router::Session
     property secret : String
     property key : String
     property expires : Int32
-    property store : Amber::Router::Cookies::Store
+    property store : Amber::Router::Cookies::SignedStore | Amber::Router::Cookies::EncryptedStore
     property session : SessionHash
 
     forward_missing_to session
@@ -43,25 +43,21 @@ module Amber::Router::Session
       session.clear
     end
 
-    def []=(key, value)
-      session[key.to_s] = value
-    end
-
     def update(hash : Hash(String | Symbol, String))
       hash.each { |key, value| session[key.to_s] = value }
       session
     end
 
     def set_session
-      store.encrypted.set(key, session.to_json, expires: expires_at, http_only: true)
+      store.set(key, session.to_json, expires: expires_at, http_only: true)
     end
 
     def expires_at
-        (Time.now + expires.seconds) if @expires > 0
+      (Time.now + expires.seconds) if @expires > 0
     end
 
     def current_session
-      SessionHash.from_json(store.encrypted[key] || "{}")
+      SessionHash.from_json(store[key] || "{}")
     rescue e : JSON::ParseException
       SessionHash.new
     end