--- a/src/amber/router/session/redis_store.cr
+++ b/src/amber/router/session/redis_store.cr
@@ -80,7 +80,7 @@ module Amber::Router::Session
     end
 
     def expires_at
-        (Time.now + expires.seconds) if @expires > 0
+      (Time.now + expires.seconds) if @expires > 0
     end
 
     def current_session