--- a/src/amber/support/message_verifier.cr
+++ b/src/amber/support/message_verifier.cr
@@ -1,4 +1,3 @@
-require "json"
 require "openssl/hmac"
 require "crypto/subtle"
 
@@ -7,22 +6,19 @@ module Amber::Support
     def initialize(@secret : Bytes, @digest = :sha1)
     end
 
-    def valid_message?(signed_message)
-      split_message = signed_message.to_s.split("--", 2)
-      return if split_message.size < 2
-      data, digest = split_message
+    def valid_message?(data, digest)
       data.size > 0 && digest.size > 0 && Crypto::Subtle.constant_time_compare(digest, generate_digest(data))
     end
 
     def verified(signed_message : String)
-      if valid_message?(signed_message)
-        begin
-          data = signed_message.split("--").first
+      begin
+        data, digest = signed_message.split("--")
+        if valid_message?(data, digest)
           String.new(decode(data))
-        rescue argument_error : ArgumentError
-          return if argument_error.message =~ %r{invalid base64}
-          raise argument_error
         end
+      rescue argument_error : ArgumentError
+        return if argument_error.message =~ %r{invalid base64}
+        raise argument_error
       end
     end