… spec/amber/router/session/cookie_store_spec.cr